+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Testing function which sets up the App with a mocked WebSerial API is working.
 - Error message is now shown if user attempts to open already in-use port, closes #254.
 - Smart scrolling when max. scrollback size is reached, auto-scroll occurs to keep the same data in view. 
+- Added specific error handling for a BufferOverrunError.
 
 ### Fixed